requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
duckduckgo-search
beautifulsoup4
requests
playwright
fastapi
uvicorn
streamlit
python-dotenv
ollama
rank_bm25

